Add sections to the Rotation documentation
